PR(REFAC): Rename `client collection {add,delete,get,update}` to `client document`
Also renames document-level methods on client.Collection interface
(Add→DocumentAdd, Get→DocumentGet, etc.) and splits cbindings into
separate collection/document files.

Move document-specific code into dedicated document*.go files for
better organization after the Document{Add,Get,Update,...} rename.
