+++ cli/document_delete.go
@@ -20,7 +20,7 @@ import (
 	"github.com/sourcenetwork/defradb/internal/identity"
 )
 
-func MakeCollectionDeleteCommand(ctx context.Context) *cobra.Command {
+func MakeDocumentDeleteCommand(ctx context.Context) *cobra.Command {
 	var argDocID string
 	var filter string
 	var cmd = &cobra.Command{
@@ -62,14 +62,14 @@ func MakeCollectionDeleteCommand(ctx context.Context) *cobra.Command {
 	}
 
 	EmbedCLIExample(ctx, cmd, "delete by docID",
-		`defradb client collection delete  --name User --docID bae-123`)
+		`defradb client document delete --collection-name User --docID bae-123`)
 
 	EmbedCLIExample(ctx, cmd, "delete by docID with identity",
-		`defradb client collection delete --name User --docID bae-123 \
+		`defradb client document delete --collection-name User --docID bae-123 \
   	-i 028d53f37a19afb9a0dbc5b4be30c65731479ee8cfa0c9bc8f8bf198cc3c075f`)
 
 	EmbedCLIExample(ctx, cmd, "delete by filter",
-		`defradb client collection delete --name User --filter '{ "_gte": { "points": 100 } }'`)
+		`defradb client document delete --collection-name User --filter '{ "_gte": { "points": 100 } }'`)
 
 	cmd.Flags().StringVar(&argDocID, "docID", "", "Document ID")
 	cmd.Flags().StringVar(&filter, "filter", "", "Document filter")
