+++ docs/website/references/cli/defradb_client_document.md
@@ -1,39 +1,26 @@
-## defradb client collection get
+## defradb client document
 
-View document fields.
+Interact with documents.
 
 ### Synopsis
 
-View document fields.
-
-```
-defradb client collection get [-i --identity] [--show-deleted] <docID>  [flags]
-```
-
-### Examples
-
-```
-Get document by ID:  
-  defradb client collection get --name User bae-123
-
-Get a private document using an identity:  
-  defradb client collection get --name User bae-123 \
-	-i 028d53f37a19afb9a0dbc5b4be30c65731479ee8cfa0c9bc8f8bf198cc3c075f 
-```
+Add, read, update, and delete documents within a collection.
 
 ### Options
 
 ```
-  -h, --help           help for get
-      --show-deleted   Show deleted documents
+      --collection-id string     Collection ID
+      --collection-name string   Collection name
+      --get-inactive             Get inactive collections as well as active
+  -h, --help                     help for document
+  -i, --identity string          Hex formatted private key used to authenticate with ACP
+      --tx uint                  Transaction ID
+      --version-id string        Collection version ID
 ```
 
 ### Options inherited from parent commands
 
 ```
-      --collection-id string        Collection ID
-      --get-inactive                Get inactive collections as well as active
-  -i, --identity string             Hex formatted private key used to authenticate with ACP
       --keyring-backend string      Keyring backend to use. Options are file or system (default "file")
       --keyring-namespace string    Service name to use when using the system backend (default "defradb")
       --keyring-path string         Path to store encrypted keys when using the file backend (default "keys")
@@ -43,18 +30,19 @@ Get a private document using an identity:
       --log-overrides string        Logger config overrides. Format <name>,<key>=<val>,...;<name>,...
       --log-source                  Include source location in logs
       --log-stacktrace              Include stacktrace in error and fatal logs
-      --name string                 Collection name
       --no-keyring                  Disable the keyring and generate ephemeral keys
       --no-log-color                Disable colored log output
       --rootdir string              Directory for persistent data (default: $HOME/.defradb)
       --secret-file string          Path to the file containing secrets (default ".env")
       --source-hub-address string   The SourceHub address authorized by the client to make SourceHub transactions on behalf of the actor
-      --tx uint                     Transaction ID
       --url string                  URL of HTTP endpoint to listen on or connect to (default "127.0.0.1:9181")
-      --version-id string           Collection version ID
 ```
 
 ### SEE ALSO
 
-* [defradb client collection](defradb_client_collection.md)	 - Interact with a collection.
+* [defradb client](defradb_client.md)	 - Interact with a DefraDB node
+* [defradb client document add](defradb_client_document_add.md)	 - Add a new document.
+* [defradb client document delete](defradb_client_document_delete.md)	 - Delete documents by docID or filter.
+* [defradb client document get](defradb_client_document_get.md)	 - View document fields.
+* [defradb client document update](defradb_client_document_update.md)	 - Update documents by docID or filter.
