+++ docs/website/references/cli/defradb_client_document_update.md
@@ -1,41 +1,46 @@
-## defradb client collection delete
+## defradb client document update
 
-Delete documents by docID or filter.
+Update documents by docID or filter.
 
 ### Synopsis
 
-Delete documents by docID or filter and lists the number of documents deleted.
+Update documents by docID or filter.
 
 ```
-defradb client collection delete [-i --identity] [--filter <filter> --docID <docID>] [flags]
+defradb client document update [-i --identity] [--filter <filter> --docID <docID>] --updater <updater> [flags]
 ```
 
 ### Examples
 
 ```
-delete by docID:  
-  defradb client collection delete  --name User --docID bae-123
+update by filter:  
+  defradb client document update --collection-name User \
+  --filter '{ "points": { "_gte": 100 } }' --updater '{ "verified": true }'
 
-delete by docID with identity:  
-  defradb client collection delete --name User --docID bae-123 \
-  	-i 028d53f37a19afb9a0dbc5b4be30c65731479ee8cfa0c9bc8f8bf198cc3c075f
+update by docID:  
+  defradb client document update --collection-name User \
+  --docID bae-123 --updater '{ "verified": true }'
 
-delete by filter:  
-  defradb client collection delete --name User --filter '{ "_gte": { "points": 100 } }'
+update private docID, with identity:  
+  defradb client document update --collection-name User \
+  -i 028d53f37a19afb9a0dbc5b4be30c65731479ee8cfa0c9bc8f8bf198cc3c075f \
+  --docID bae-123 --updater '{ "verified": true }'
 ```
 
 ### Options
 
 ```
-      --docID string    Document ID
-      --filter string   Document filter
-  -h, --help            help for delete
+      --docID string     Document ID
+      --filter string    Document filter
+  -h, --help             help for update
+      --updater string   Document updater
 ```
 
 ### Options inherited from parent commands
 
 ```
       --collection-id string        Collection ID
+      --collection-name string      Collection name
       --get-inactive                Get inactive collections as well as active
   -i, --identity string             Hex formatted private key used to authenticate with ACP
       --keyring-backend string      Keyring backend to use. Options are file or system (default "file")
@@ -47,7 +52,6 @@ delete by filter:
       --log-overrides string        Logger config overrides. Format <name>,<key>=<val>,...;<name>,...
       --log-source                  Include source location in logs
       --log-stacktrace              Include stacktrace in error and fatal logs
-      --name string                 Collection name
       --no-keyring                  Disable the keyring and generate ephemeral keys
       --no-log-color                Disable colored log output
       --rootdir string              Directory for persistent data (default: $HOME/.defradb)
@@ -60,5 +64,5 @@ delete by filter:
 
 ### SEE ALSO
 
-* [defradb client collection](defradb_client_collection.md)	 - Interact with a collection.
+* [defradb client document](defradb_client_document.md)	 - Interact with documents.
